git: make GitBackend colocation-aware
GitBackend now knows if it's colocated or not. This lays the foundation
for GitBackend opening the worktree version of a git repo.

Later in these diffs, we add worktree support. Then, automatically, all
HEADs will be written to those worktrees instead. We won't have to add
special code to re-open the git repo at the worktree (= workspace root)
in a dozen places in the CLI, and then maintain discipline in selecting
which repo to use when.